feat: :sparkles: construct and export separate main and renderer clients
